+++ auto_py_to_exe/web/js/initialise.js
@@ -67,6 +67,9 @@ window.addEventListener("load", async () => {
     // Setup ui events (for static content) and setup initial state
     setupEvents();
 
+    // Setup language selection
+    setupLanguageSelection();
+
     // Setup advanced section (for dynamic content)
     constructAdvancedSection(options);
 
@@ -88,6 +91,9 @@ window.addEventListener("load", async () => {
 
     // Update the current command when setup is complete
     updateCurrentCommandDisplay();
+
+    // Try to translate to the default browser language
+    translate();
 
     // If the server stops, close the UI
     window.eel._websocket.addEventListener('close', e => window.close());